Fin modélisation 3D

+++ database/migrations/2021_03_14_173145_create_projet_plantes_3d.php
@@ -19,9 +19,9 @@ class CreateProjetPlantes3d extends Migration
             $table->integer('coordx');
             $table->integer('coordy');
             $table->integer('coordz');
-            $table->integer('rotationx');
-            $table->integer('rotationy');
-            $table->integer('rotationz');
+            $table->float('rotationx');
+            $table->float('rotationy');
+            $table->float('rotationz');
         });
     }
